pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "zotero-bridge"
version = "1.0.0"
description = "Public REST bridge between ChatGPT/Codex and Zotero Web API v3"
readme = "README_IMPLEMENTATION.md"
requires-python = ">=3.12"
dependencies = [
  "bleach>=6.2.0,<7",
  "fastapi>=0.115.0,<1",
  "httpx>=0.28.0,<1",
  "markdown-it-py>=3.0.0,<4",
  "pydantic-settings>=2.8.0,<3",
  "pydantic>=2.11.0,<3",
  "pypdf>=6.7.5",
  "python-multipart>=0.0.20,<1",
  "uvicorn>=0.34.0,<1",
]

[dependency-groups]
dev = [
  "mypy>=1.15.0,<2",
  "pyyaml>=6.0.2,<7",
  "pytest>=8.3.0,<9",
  "pytest-asyncio>=0.26.0,<1",
  "respx>=0.22.0,<1",
  "ruff>=0.11.0,<1",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "B",
  "UP",
  "ASYNC",
]

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_equality = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
plugins = []
packages = ["app", "tests"]

[[tool.mypy.overrides]]
module = ["bleach", "bleach.*", "yaml", "yaml.*"]
ignore_missing_imports = true
